Fix toggleBounce, took off click handler, invoked function directly. Line 294
Andrew pointed out :
Consider this: on initialization of the ViewModel, the marker click handlers are already automatically getting made in the course of processing the allPlaces array. In other words, the click handlers already exist. Inside that handler function you want toggleBounce to be invoked instead of attached since we are already reacting to a click event!

The event has already happened by that point we are calling getPlacesDetails.